Added wpe2k-fs-monitor.sh
This will prompt the user when running the main script if they'd like to create a service. This service will watch for windows in full screen or using all width and 95% height. If it finds that to be the case it will kill komorebi completely. The service also checks for multiple instances of komorebi running. This will prevent the system from flooding.